refactor(today): extract recommendation builders
